+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageView.swift
@@ -252,7 +252,8 @@ struct StreamTextView: View {
 public struct LinkDetectionTextView: View {
     @Environment(\.layoutDirection) var layoutDirection
     @Environment(\.channelTranslationLanguage) var translationLanguage
-    
+    @Environment(\.messageViewModel) var messageViewModel
+
     @Injected(\.colors) var colors
     @Injected(\.fonts) var fonts
     @Injected(\.utils) var utils
@@ -262,43 +263,36 @@ public struct LinkDetectionTextView: View {
     var text: LocalizedStringKey {
         LocalizedStringKey(message.adjustedText)
     }
-    
-    @State var displayedText: AttributedString?
-    
+
+    // The translations store is used to detect changes so the textContent is re-rendered.
+    // The @Environment(\.messageViewModel) is not reactive like @EnvironmentObject.
+    // TODO: On v5 the TextView should be refactored and not depend directly on the view model.
+    @ObservedObject var originalTranslationsStore = InjectedValues[\.utils].originalTranslationsStore
+
     @State var linkDetector = TextLinkDetector()
-    
     @State var tintColor = InjectedValues[\.colors].tintColor
         
-    public init(message: ChatMessage) {
+    public init(
+        message: ChatMessage
+    ) {
         self.message = message
     }
     
     public var body: some View {
         Group {
-            if let displayedText {
-                Text(displayedText)
-            } else {
-                Text(message.adjustedText)
-            }
+            Text(displayText)
         }
         .foregroundColor(textColor(for: message))
         .font(fonts.body)
         .tint(tintColor)
-        .onAppear {
-            displayedText = attributedString(for: message)
+        .onChange(of: message) { message in
+            messageViewModel?.message = message
         }
-        .onChange(of: message, perform: { updated in
-            displayedText = attributedString(for: updated)
-        })
     }
     
-    private func attributedString(for message: ChatMessage) -> AttributedString {
-        var text = message.adjustedText
-        
-        // Translation
-        if let translatedText = message.textContent(for: translationLanguage) {
-            text = translatedText
-        }
+    var displayText: AttributedString {
+        let text = messageViewModel?.textContent ?? message.text
+
         // Markdown
         let attributes = AttributeContainer()
             .foregroundColor(textColor(for: message))