+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageContainerView_Tests.swift
@@ -5,6 +5,7 @@
 import SnapshotTesting
 @testable import StreamChat
 @testable import StreamChatSwiftUI
+@testable import StreamChatTestTools
 import StreamSwiftTestHelpers
 import SwiftUI
 import XCTest
@@ -123,7 +124,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
         )
 
         // When
-        let view = testMessageViewContainer(message: message)
+        let view = testMessageViewContainer(message: message, channel: .mockNonDMChannel())
 
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
@@ -139,7 +140,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
         )
 
         // When
-        let view = testMessageViewContainer(message: message)
+        let view = testMessageViewContainer(message: message, channel: .mockNonDMChannel())
 
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
@@ -160,7 +161,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
         )
 
         // When
-        let view = testMessageViewContainer(message: message)
+        let view = testMessageViewContainer(message: message, channel: .mockNonDMChannel())
 
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
@@ -274,7 +275,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
         )
 
         // When
-        let view = testMessageViewContainer(message: message)
+        let view = testMessageViewContainer(message: message, channel: .mockNonDMChannel())
 
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
@@ -292,7 +293,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
         )
 
         // When
-        let view = testMessageViewContainer(message: message)
+        let view = testMessageViewContainer(message: message, channel: .mockNonDMChannel())
 
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
@@ -315,7 +316,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
             .environment(\.channelTranslationLanguage, .spanish)
 
         // Then
-        AssertSnapshot(view, size: CGSize(width: 375, height: 200))
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
     }
     
     func test_translatedText_myMessageIsNotTranslated_snapshot() {
@@ -403,7 +404,7 @@ class MessageContainerView_Tests: StreamChatTestCase {
         waitForExpectations(timeout: 1)
     }
 
-    func test_isSwipeToReplyPossible_whenCanQuoteReply_whenMessageInteractable_shouldBeTrue() {
+    func test_isSwipeToReplyPossible_whenRepliesEnabled_whenMessageInteractable_shouldBeTrue() {
         let message = ChatMessage.mock(
             id: .unique,
             cid: .unique,
@@ -412,23 +413,12 @@ class MessageContainerView_Tests: StreamChatTestCase {
             isSentByCurrentUser: true
         )
 
-        let view = MessageContainerView(
-            factory: DefaultViewFactory.shared,
-            channel: .mockDMChannel(ownCapabilities: [.quoteMessage]),
-            message: message,
-            width: defaultScreenSize.width,
-            showsAllInfo: true,
-            isInThread: false,
-            isLast: false,
-            scrolledId: .constant(nil),
-            quotedMessage: .constant(nil),
-            onLongPress: { _ in }
-        )
+        let viewModel = MessageViewModel(message: message, channel: .mockDMChannel(config: .mock(repliesEnabled: true)))
 
-        XCTAssertTrue(view.isSwipeToReplyPossible)
+        XCTAssertTrue(viewModel.isSwipeToQuoteReplyPossible)
     }
 
-    func test_isSwipeToReplyPossible_whenCanNotQuoteReply_whenMessageInteractable_shouldBeFalse() {
+    func test_isSwipeToReplyPossible_whenRepliesDisabled_whenMessageInteractable_shouldBeFalse() {
         let message = ChatMessage.mock(
             id: .unique,
             cid: .unique,
@@ -437,23 +427,12 @@ class MessageContainerView_Tests: StreamChatTestCase {
             isSentByCurrentUser: true
         )
 
-        let view = MessageContainerView(
-            factory: DefaultViewFactory.shared,
-            channel: .mockDMChannel(ownCapabilities: []),
-            message: message,
-            width: defaultScreenSize.width,
-            showsAllInfo: true,
-            isInThread: false,
-            isLast: false,
-            scrolledId: .constant(nil),
-            quotedMessage: .constant(nil),
-            onLongPress: { _ in }
-        )
+        let viewModel = MessageViewModel(message: message, channel: .mockDMChannel(config: .mock(repliesEnabled: false)))
 
-        XCTAssertFalse(view.isSwipeToReplyPossible)
+        XCTAssertFalse(viewModel.isSwipeToQuoteReplyPossible)
     }
 
-    func test_isSwipeToReplyPossible_whenCanQuoteReply_whenMessageNotInteractable_shouldBeFalse() {
+    func test_isSwipeToReplyPossible_whenRepliesEnabled_whenMessageNotInteractable_shouldBeFalse() {
         let message = ChatMessage.mock(
             id: .unique,
             cid: .unique,
@@ -463,28 +442,125 @@ class MessageContainerView_Tests: StreamChatTestCase {
             isSentByCurrentUser: true
         )
 
-        let view = MessageContainerView(
-            factory: DefaultViewFactory.shared,
-            channel: .mockDMChannel(ownCapabilities: [.quoteMessage]),
-            message: message,
-            width: defaultScreenSize.width,
-            showsAllInfo: true,
-            isInThread: false,
-            isLast: false,
-            scrolledId: .constant(nil),
-            quotedMessage: .constant(nil),
-            onLongPress: { _ in }
-        )
+        let viewModel = MessageViewModel(message: message, channel: .mockDMChannel(config: .mock(repliesEnabled: true)))
 
-        XCTAssertFalse(view.isSwipeToReplyPossible)
+        XCTAssertFalse(viewModel.isSwipeToQuoteReplyPossible)
+    }
+
+    func test_translatedText_showOriginalTranslatedButtonEnabled_originalTextShown_snapshot() {
+        // Given
+        let message = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Hello",
+            author: .mock(id: .unique),
+            translations: [
+                .spanish: "Hola"
+            ]
+        )
+        let utils = Utils(
+            dateFormatter: EmptyDateFormatter(),
+            messageListConfig: .init(
+                messageDisplayOptions: MessageDisplayOptions(showOriginalTranslatedButton: true)
+            )
+        )
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+
+        // When
+        let messageViewModel = MessageViewModel_Mock(
+            message: message,
+            channel: .mock(
+                cid: .unique,
+                membership: .mock(id: .unique, language: .spanish)
+            )
+        )
+        messageViewModel.mockOriginalTextShown = true
+        let view = testMessageViewContainer(message: message, messageViewModel: messageViewModel)
+            .environment(\.channelTranslationLanguage, .spanish)
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+
+    func test_translatedText_showOriginalTranslatedButtonEnabled_translatedTextShown_snapshot() {
+        // Given
+        let message = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Hello",
+            author: .mock(id: .unique),
+            translations: [
+                .spanish: "Hola"
+            ]
+        )
+        let utils = Utils(
+            dateFormatter: EmptyDateFormatter(),
+            messageListConfig: .init(
+                messageDisplayOptions: MessageDisplayOptions(showOriginalTranslatedButton: true)
+            )
+        )
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+
+        // When
+        let messageViewModel = MessageViewModel_Mock(
+            message: message,
+            channel: .mock(
+                cid: .unique,
+                membership: .mock(id: .unique, language: .spanish)
+            )
+        )
+        messageViewModel.mockOriginalTextShown = false
+        let view = testMessageViewContainer(message: message, messageViewModel: messageViewModel)
+            .environment(\.channelTranslationLanguage, .spanish)
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+
+    func test_translatedText_showOriginalTranslatedButtonDisabled_translatedTextShown_snapshot() {
+        // Given
+        let message = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Hello",
+            author: .mock(id: .unique),
+            translations: [
+                .spanish: "Hola"
+            ]
+        )
+        let utils = Utils(
+            dateFormatter: EmptyDateFormatter(),
+            messageListConfig: .init(
+                messageDisplayOptions: MessageDisplayOptions(showOriginalTranslatedButton: false)
+            )
+        )
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+
+        // When
+        let messageViewModel = MessageViewModel_Mock(
+            message: message,
+            channel: .mock(
+                cid: .unique,
+                membership: .mock(id: .unique, language: .spanish)
+            )
+        )
+        let view = testMessageViewContainer(message: message, messageViewModel: messageViewModel)
+            .environment(\.channelTranslationLanguage, .spanish)
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
     }
 
     // MARK: - private
 
-    func testMessageViewContainer(message: ChatMessage) -> some View {
+    func testMessageViewContainer(
+        message: ChatMessage,
+        channel: ChatChannel? = nil,
+        messageViewModel: MessageViewModel? = nil
+    ) -> some View {
         MessageContainerView(
             factory: DefaultViewFactory.shared,
-            channel: .mockDMChannel(),
+            channel: channel ?? .mockDMChannel(),
             message: message,
             width: defaultScreenSize.width,
             showsAllInfo: true,
@@ -492,8 +568,17 @@ class MessageContainerView_Tests: StreamChatTestCase {
             isLast: false,
             scrolledId: .constant(nil),
             quotedMessage: .constant(nil),
-            onLongPress: { _ in }
+            onLongPress: { _ in },
+            viewModel: messageViewModel ?? MessageViewModel(message: message, channel: channel)
         )
         .frame(width: 375, height: 200)
+    }
+}
+
+class MessageViewModel_Mock: MessageViewModel {
+    var mockOriginalTextShown: Bool = false
+
+    override var originalTextShown: Bool {
+        mockOriginalTextShown
     }
 }